+++ scripts/train.py
@@ -57,7 +57,7 @@ class Args:
     """if toggled, the environment will return visual observation otherwise it would not"""
     shaping_reward: bool = True
     """if toggled, the environment will utilize dense shaping reward in training otherwise it would only use sparse goal"""
-    sparse_reward: int = 10.0
+    sparse_reward: float = 10.0
     """total sparse reward for success"""
 
     # Algorithm specific arguments
